requirements.txt
flask==2.3.3
werkzeug==2.3.7
flask-sqlalchemy==3.0.2
flask-login==0.6.2
flask-migrate==3.1.0
python-dotenv==0.19.0
Pillow>=10.0.0
numpy<2.0.0
torch==2.2.0
torchvision==0.17.0
google-generativeai==0.3.2
